adds sane error messages
Well, at least they are more sane than were before :)

1) we track all types that were used during unification and dump them
to users in hope that this will clarify them the issue

2) we disallow unresolved functions and variables and immediately
lower the type to bottom if we see one.

3) the error messages are now more explonatory and include source code
with the relevenat context and highlighting.

+++ plugins/primus_lisp/primus_lisp_io.ml
@@ -264,7 +264,7 @@ let init redirections =
             channel that has the descriptor DESCR to be outputted to the
             associated destination. Returns -1 if no such channel exists or
             if in case of an IO error.|};
-        def "channel-input"  (one int @-> int) (module Input)
+        def "channel-input"  (one int @-> byte) (module Input)
           {|(channel-input DESC) reads one byte from a channel that
             has the descriptor DESC. Returns -1 if no such channel
             exists, or if any IO error occurs, if the channel is not
